fix: Authorized access to endpoint based on roles and fixed all unit test to pass

+++ backend/src/main/java/dev/levelupschool/backend/service/implementation/LevelUpArticleService.java
@@ -7,16 +7,19 @@ import dev.levelupschool.backend.data.model.Article;
 import dev.levelupschool.backend.data.model.User;
 import dev.levelupschool.backend.data.repository.ArticleRepository;
 import dev.levelupschool.backend.exception.ModelNotFoundException;
+import dev.levelupschool.backend.exception.UserException;
 import dev.levelupschool.backend.service.interfaces.ArticleService;
 import dev.levelupschool.backend.service.interfaces.UserService;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class LevelUpArticleService implements ArticleService {
     private final ArticleRepository articleRepository;
     private final UserService userService;
+
     public LevelUpArticleService(
         ArticleRepository articleRepository,
         UserService userService){
@@ -24,8 +27,8 @@ public class LevelUpArticleService implements ArticleService {
         this.userService = userService;
     }
     @Override
-    public CreateArticleResponse createArticle(CreateArticleRequest createArticleRequest) {
-        User fondUser = userService.findUserById(createArticleRequest.getAuthorId());
+    public CreateArticleResponse createArticle(CreateArticleRequest createArticleRequest, String authHeader) {
+        User fondUser = userService.getUser(authHeader);
         Article newArticle = new Article(
             createArticleRequest.getTitle(),
             createArticleRequest.getContent(),
@@ -54,10 +57,12 @@ public class LevelUpArticleService implements ArticleService {
     }
 
     @Override
-    public Article updateArticle(UpdateArticleRequest updateArticleRequest, Long articleId) {
+    public Article updateArticle(UpdateArticleRequest updateArticleRequest, Long articleId, String authHeader) {
+        User foundUser = userService.getUser(authHeader);
         return articleRepository
             .findById(articleId)
             .map(article -> {
+                if (!Objects.equals(foundUser.getId(), article.getUser().getId()))throw new UserException("You have no permission to update article");
                 article.setTitle(updateArticleRequest.getTitle());
                 article.setContent(updateArticleRequest.getContent());
                 return articleRepository.save(article);
@@ -66,7 +71,11 @@ public class LevelUpArticleService implements ArticleService {
     }
 
     @Override
-    public void deleteArticle(Long articleId) {
+    public void deleteArticle(Long articleId, String authHeader) {
+        Article foundArticle = articleRepository.findById(articleId)
+            .orElseThrow(()-> new ModelNotFoundException(articleId));
+        User foundUser = userService.getUser(authHeader);
+        if (!Objects.equals(foundUser.getId(), foundArticle.getUser().getId()))throw new UserException("You have no permission to delete article");
         articleRepository
             .deleteById(articleId);
     }
