+++ backend/src/main/java/dev/levelupschool/backend/service/implementation/LevelUpCommentService.java
@@ -8,12 +8,14 @@ import dev.levelupschool.backend.data.model.User;
 import dev.levelupschool.backend.data.model.Comment;
 import dev.levelupschool.backend.data.repository.CommentRepository;
 import dev.levelupschool.backend.exception.ModelNotFoundException;
+import dev.levelupschool.backend.exception.UserException;
 import dev.levelupschool.backend.service.interfaces.ArticleService;
 import dev.levelupschool.backend.service.interfaces.UserService;
 import dev.levelupschool.backend.service.interfaces.CommentService;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class LevelUpCommentService implements CommentService {
@@ -30,8 +32,8 @@ public class LevelUpCommentService implements CommentService {
     }
 
     @Override
-    public AddCommentResponse addComment(AddCommentRequest addCommentRequest) {
-        User fondUser = userService.findUserById(addCommentRequest.getAuthorId());
+    public AddCommentResponse addComment(AddCommentRequest addCommentRequest, String authHeader) {
+        User fondUser = userService.getUser(authHeader);
         Article foundArticle = articleService.findArticleById(addCommentRequest.getArticleId());
         Comment newComment = new Comment(addCommentRequest.getContent(), foundArticle, fondUser);
         Comment savedComment = commentRepository.save(newComment);
@@ -56,10 +58,12 @@ public class LevelUpCommentService implements CommentService {
     }
 
     @Override
-    public Comment updateComment(UpdateCommentRequest updateCommentRequest, Long commentId) {
+    public Comment updateComment(UpdateCommentRequest updateCommentRequest, Long commentId, String authHeader) {
+        User foundUser = userService.getUser(authHeader);
         return commentRepository
             .findById(commentId)
             .map(comment -> {
+                if (!Objects.equals(foundUser.getId(), comment.getUser().getId()))throw new UserException("You have no permission to update comment");
                 comment.setContent(updateCommentRequest.getContent());
                 return commentRepository.save(comment);
             })
@@ -67,7 +71,11 @@ public class LevelUpCommentService implements CommentService {
     }
 
     @Override
-    public void deleteComment(Long commentId) {
+    public void deleteComment(Long commentId, String authHeader) {
+        User foundUser = userService.getUser(authHeader);
+        Comment foundComment = commentRepository.findById(commentId)
+                .orElseThrow(()-> new ModelNotFoundException(commentId));
+        if (!Objects.equals(foundUser.getId(), foundComment.getUser().getId()))throw new UserException("You have no permission to update comment");
         commentRepository
             .deleteById(commentId);
     }
